third_party/WebKit/LayoutTests/fast/canvas/gradient-addColorStop-with-invalid-color-expected.txt

This test checks invalid colors on gradients.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS gradient.addColorStop(0, '') threw exception SyntaxError: Failed to execute 'addColorStop' on 'CanvasGradient': The value provided ('') could not be parsed as a color..
PASS gradient.addColorStop(0, '#cc') threw exception SyntaxError: Failed to execute 'addColorStop' on 'CanvasGradient': The value provided ('#cc') could not be parsed as a color..
PASS gradient.addColorStop(0, 'rgb(257, 0)') threw exception SyntaxError: Failed to execute 'addColorStop' on 'CanvasGradient': The value provided ('rgb(257, 0)') could not be parsed as a color..
PASS gradient.addColorStop(0, 'rgb(257, 0, 5, 0)') threw exception SyntaxError: Failed to execute 'addColorStop' on 'CanvasGradient': The value provided ('rgb(257, 0, 5, 0)') could not be parsed as a color..
PASS successfullyParsed is true

TEST COMPLETE
